@@ -3,8 +3,9 @@ package com.github.frimtec.android.pikettassist.state;
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.support.annotation.Nullable;
 import android.util.Log;
+
+import androidx.annotation.Nullable;
 
 import static com.github.frimtec.android.pikettassist.domain.OnOffState.OFF;
 
